버그 수정

+++ StereoMix-Launcher/src/FileHelper.cs
@@ -169,7 +169,10 @@ public static class FileHelper
                 await HttpHelper.SaveToFile(window, response, fileStream);
             });
             
-            Directory.Delete(window.InstallDirectory, true);
+            if (Directory.Exists(window.InstallDirectory))
+            {
+                Directory.Delete(window.InstallDirectory, true);
+            }
             
             ZipFile.ExtractToDirectory(tempFile, window.InstallDirectory, true);
             File.Delete(tempFile);
